PoissonRandomVariable object

Defines a Poisson random variable. It uses in the backend all functions that defines a Poisson random variable. As the other definitions in distributions this is an R6 object. Tha parametrization used is from a $$X \sim \mathsf{Poisson}(\lambda)\,,$$ where the parameter \(\lambda >0\) denotes the rate of ocurrence of events.

Note

All random variables in distributions are defined as R6 objects. The fields (referenced below) are needed for objects of R6::R6Class. In our context these are what defines which specific instance of a random variable is used. That is, the public fields are the parameters of the random variable \(\theta \in \mathbb{R}^p\), for some \(p\).

See also

Super class

distributions::DiscreteRandomVariable -> Poisson

Public fields

lambda

(double, positive)
the number of independent trials, \(\lambda > 0\).

Methods

Public methods

PoissonRandomVariable$new()

PoissonRandomVariable$sample()

PoissonRandomVariable$density()

Method new()

Generates a Binomial DiscreteRandomVariable object with specified number of trials (n) and probability of success (theta).

Usage

PoissonRandomVariable$new(lambda = 1)

Arguments

lambda

(double, positive) the rate of events.

Method sample()

Generates random variables using the stats::rpois() function.

Usage

PoissonRandomVariable$sample(nsamples = 1)

Arguments

nsamples

(integer) the number of random numbers being generated that behaves like the random variables.

Returns

(array) of random numbers from the distribution.

Method density()

Evaluates the density function using the stats::dpois() function.

Usage

PoissonRandomVariable$density(x, log = TRUE)

Arguments

x

(array) to evaluate the density function.

log

(Boolean) indicates if we are evaluating the log-density.

Returns

(array) of evaluations at x.