Create a binomial distribution

Bernoulli distributions are used to represent situations can that can thought of as size (often called \(n\) in textbooks) independent coin flips, where each coin flip has probability p of success. The bernoulli() distribution is a special case of binomial distribution when n = 1.

binomial(size, p)

Arguments

size
The number of trials. Must be an integer greater than or equal to one. When size = 1L, the binomial distribution reduces to the bernoulli distribution. Oftened called n in textbooks.

p
The success probability for a given trial. p can be any value in [0, 1], and defaults to 0.5. greater than zero. Defaults to 1.

Value

A binomial object.

Details

The binomial distribution comes up when you are interested in the portion of people who do a thing. The binomial distribution also comes up in the sign test, sometimes called the binomial test (see stats::binom.test()), where you may need the binomial C.D.F. to compute p-values.

We recommend reading this documentation on https://alexpghayes.github.io/distributions, where the math will render with additional detail.

In the following, let \(X\) be a Binomial random variable with parameter size = \(n\) and p = \(p\). Some textbooks define \(q = 1 - p\), or called \(\pi\) instead of \(p\).

Support: \(\{0, 1, 2, ..., n\}\)

Mean: \(np\)

Variance: \(np \cdot (1 - p) = np \cdot q\)

Probability mass function (p.m.f):

$$ P(X = k) = \binom{n}{k} p^k (1 - p)^{n-k} $$

Cumulative distribution function (c.d.f):

$$ P(X \le k) = \sum_{i=0}^{\lfloor k \rfloor} \binom{n}{i} p^i (1 - p)^{n-i} $$

Moment generating function (m.g.f):

$$ \mathbb{E}(e^{tX}) = (1 - p + p e^t)^n $$

See also

Examples

b <- binomial(10, 0.2)
b

#> Binomial distribution (size = 10, p = 0.2)

random(b, 10)

#>  [1] 3 4 1 1 3 2 2 2 0 3

pdf(b, 2L)

#> [1] 0.3019899

cdf(b, 4L)

#> [1] 0.9672065

quantile(b, 0.7)

#> [1] 3

cdf(b, quantile(b, 0.7))

#> [1] 0.8791261

quantile(b, cdf(b, 7))

#> [1] 7