Create a chi square distribution

Chi square distributions show up often in frequentist settings as the sampling distribution of test statistics, especially in maximum likelihood estimation settings. The chi square distribution with a single degree of freedom is equivalent to a squared standard normal distribution. Summing chi square distributions results in a new chi square distribution with added degrees of freedom. A chi square distribution is a special case of the gamma distribution with shape = 2 (double check this). TODO: add relationship to F and T distributions.

chi_square(df)

Arguments

df
Degrees of freedom. Must be positive.

Value

A chi_square object.

Details

We recommend reading this documentation on https://alexpghayes.github.io/distributions, where the math will render with additional detail and much greater clarity.

In the following, let \(X\) be a \(\chi^2\) random variable with df = \(k\).

Support: \(\mathbb{R}^+\), the set of positive real numbers

Mean: \(k\)

Variance: \(2k\)

Probability density function (p.d.f):

$$ f(x) = \frac{1}{\sqrt{2 \pi \sigma^2}} e^{-(x - \mu)^2 / 2 \sigma^2} $$

Cumulative distribution function (c.d.f):

The cumulative distribution function has the form

$$ F(t) = \int_{-\infty}^t \frac{1}{\sqrt{2 \pi \sigma^2}} e^{-(x - \mu)^2 / 2 \sigma^2} dx $$

but this integral does not have a closed form solution and must be approximated numerically. The c.d.f. of a standard normal is sometimes called the "error function". The notation \(\Phi(t)\) also stands for the c.d.f. of a standard normal evaluated at \(t\). Z-tables list the value of \(\Phi(t)\) for various \(t\).

Moment generating function (m.g.f):

$$ \mathbb{E}(e^{tX}) = e^{\mu t + \sigma^2 t^2 / 2} $$

See also

Examples

c <- chi_square(5)
c

#> chi square distribution (df = 5)

random(c, 10)

#>  [1] 11.2129049  7.8935724  2.1298341  5.2084236  5.4563211  3.6636712
#>  [7] 10.9823299  0.7858347  4.8748588  1.7938110

pdf(c, 2)

#> [1] 0.1383692

cdf(c, 4)

#> [1] 0.450584

quantile(c, 0.7)

#> [1] 6.06443

cdf(c, quantile(c, 0.7))

#> [1] 0.7

quantile(c, cdf(c, 7))

#> [1] 7