Determine quantiles of a normal distribution

Please see the documentation of normal() for some properties of the normal distribution, as well as extensive examples showing to how calculate p-values and confidence intervals. quantile()

# S3 method for normal
quantile(d, p, ...)

Arguments

d
A normal object created by a call to normal().

p
A vector of probabilites.

...
Unused. Unevaluated arguments will generate a warning to catch mispellings or other possible errors.

Value

A vector of quantiles, one for each element of p.

Details

This function returns the same values that you get from a Z-table. Note quantile() is the inverse of cdf(). Please see the documentation of normal() for some properties of the normal distribution, as well as extensive examples showing to how calculate p-values and confidence intervals.

See also

Examples

n <- normal(5, 2)
n

#> normal distribution (mu = 5, sigma = 2)

random(n, 10)

#>  [1] 4.426760 7.441544 6.601643 4.002187 5.883946 8.267753 4.737248 5.336119
#>  [9] 7.033873 5.226213

pdf(n, 2)

#> [1] 0.0647588

cdf(n, 4)

#> [1] 0.3085375

quantile(n, 0.7)

#> [1] 6.048801

### example: calculating p-values for two-sided Z-test

# here the null hypothesis is H_0: mu = 3
# and we assume sigma = 2

# exactly the same as: Z <- normal(0, 1)
Z <- normal()

# data to test
x <- c(3, 7, 11, 0, 7, 0, 4, 5, 6, 2)
nx <- length(x)

# calculate the z-statistic
z_stat <- (mean(x) - 3) / (2 / sqrt(nx))
z_stat

#> [1] 2.371708

# calculate the two-sided p-value
1 - cdf(Z, abs(z_stat)) + cdf(Z, -abs(z_stat))

#> [1] 0.01770607

# exactly equivalent to the above
2 * cdf(Z, -abs(z_stat))

#> [1] 0.01770607

# p-value for one-sided test
# H_0: mu <= 3   vs   H_A: mu > 3
1 - cdf(Z, z_stat)

#> [1] 0.008853033

# p-value for one-sided test
# H_0: mu >= 3   vs   H_A: mu < 3
cdf(Z, z_stat)

#> [1] 0.991147

### example: calculating a 88 percent Z CI for a mean

# same `x` as before, still assume `sigma = 2`

# lower-bound
mean(x) - quantile(Z, 1 - 0.12 / 2) * 2 / sqrt(nx)

#> [1] 3.516675

# upper-bound
mean(x) + quantile(Z, 1 - 0.12 / 2) * 2 / sqrt(nx)

#> [1] 5.483325

# equivalent to
mean(x) + c(-1, 1) * quantile(Z, 1 - 0.12 / 2) * 2 / sqrt(nx)

#> [1] 3.516675 5.483325

# also equivalent to
mean(x) + quantile(Z, 0.12 / 2) * 2 / sqrt(nx)

#> [1] 3.516675

mean(x) + quantile(Z, 1 - 0.12 / 2) * 2 / sqrt(nx)

#> [1] 5.483325

### generating random samples and plugging in ks.test()

set.seed(27)

# generate a random sample
normal_sample <- random(normal(3, 7), 26)

# test if sample is normal(3, 7)
ks.test(normal_sample, pnorm, mean = 3, sd = 7)

#> 
#> 	One-sample Kolmogorov-Smirnov test
#> 
#> data:  normal_sample
#> D = 0.20352, p-value = 0.2019
#> alternative hypothesis: two-sided
#> 

# test if sample is gamma(8, 3) using base R pgamma()
ks.test(normal_sample, pgamma, shape = 8, rate = 3)

#> 
#> 	One-sample Kolmogorov-Smirnov test
#> 
#> data:  normal_sample
#> D = 0.46154, p-value = 1.37e-05
#> alternative hypothesis: two-sided
#> 

###

# note that the cdf() and quantile() functions are inverses
cdf(n, quantile(n, 0.7))

#> [1] 0.7

quantile(n, cdf(n, 7))

#> [1] 7